P9_Modelisez_une_infrastructure_dans_le_cloud_aymeric_bailleul/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.3.3
pyspark==3.5.1

# Development tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
